# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Food Court QR Ordering Platform API
  version: 1.0.0-pilot-STALE
  description: |
    # ⚠ THIS DOCUMENT IS STALE. DO NOT GENERATE A CLIENT FROM IT.

    Superseded by `docs/prd.md` v7.0 and by the running code. It is kept only so
    the divergence is visible; every disagreement below resolves in favour of
    the implementation.

    The warning lives here, in the file itself, rather than only in the PRD.
    Somebody who opens this directly — to generate a client, to check a field
    name, to write an integration — would otherwise get no warning at all, and
    a stale contract read in good faith produces confidently wrong code.

    ## What is known to be wrong

    | Here | Actually |
    | --- | --- |
    | `/qr/{token}` returns a `table` object | It resolves to a FOOD COURT. The table concept was removed — PRD §2.1 |
    | Order status `COMPLETED` | `COLLECTED`. Cooked and handed over are different events — PRD §7.1 |
    | Payment status `SUCCESS` | `AUTHORIZED` and `CAPTURED` are separate states — PRD §8.1 |
    | Menu item `is_available` boolean | `status` / `availability` / `inventory_mode`, three concepts — PRD §6 |
    | `POST /kds/orders/{id}/complete` | `/collect` |
    | no session or OTP endpoints | `/auth/otp/request`, `/auth/otp/verify` exist — PRD §11.2 |
    | no payment endpoints | `/orders/{id}/payment-intent`, `/webhooks/payments`, `/orders/{id}/payment` exist |
...

...
    script produces types that contradict the server.

...

...

...

...

...
servers:
...
security:
  - sessionToken: []
paths:
...
  /food-courts/{foodCourtId}/vendors:
    get:
      tags: [Discovery]
      operationId: listVendors
      summary: List vendors in the active food court
      description: |
        PRD CUS-VEN-01/02/03. `orderable` is derived at read time from
        configured hours, temporary closure **and KDS heartbeat freshness**.
        A vendor whose device has been silent past the threshold is returned
        with `orderable: false` and `unavailableReason: DEVICE_OFFLINE`.
      parameters:
        - name: foodCourtId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Vendor list
          content:
            application/json:
              schema:
                type: object
                properties:
                  vendors:
                    type: array
                    items: { $ref: '#/components/schemas/VendorSummary' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /manager/orders/{orderId}/force-cancel:
    post:
      tags: [Manager]
      operationId: forceCancelOrder
      description: PRD MGR-03. Reason mandatory. Audited. Triggers the refund command.
      parameters:
        - { $ref: '#/components/parameters/OrderId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 3, maxLength: 300 }
      responses:
        '200':
          description: Cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
...
  /admin/vendors:
    post:
      tags: [Admin]
      operationId: createVendor
      description: |
        PRD VEN-ONB-01/02. `settlementMode` is required before a vendor can be
        activated, and a `PLATFORM_COLLECT` vendor cannot be activated without a
        linked sub-merchant account.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VendorInput' }
      responses:
        '201':
          description: Vendor created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vendor' }
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived customer session token issued by POST /sessions.
    deviceToken:
      type: http
      scheme: bearer
      description: KDS or printer device token. Scoped to one vendor.
    staffToken:
      type: http
      scheme: bearer
      description: Manager, ops, finance and admin. Role and tenant scope are claims.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated, unique per attempt. Backed by a unique database
        constraint. Replay returns the original result. PRD API-01.
      schema: { type: string, minLength: 16, maxLength: 128 }
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    NotFound:
      description: |
        Not found, or not visible to this tenant. Cross-tenant reads return 404
        rather than 403 so existence is not leaked (PRD RBAC-01).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [code, message, correlationId]
      properties:
        code:
          type: string
          description: |
            Stable machine-readable error code. Clients branch on this, never on
            the message. See TDD §18 for the full catalogue.
          examples:
            [VENDOR_DEVICE_OFFLINE, INVALID_TRANSITION, QUOTE_EXPIRED, IDEMPOTENCY_KEY_REQUIRED]
        message: { type: string, description: Human-readable, not for display to customers. }
        detail: { type: string, nullable: true }
        correlationId: { type: string }
        fields:
          type: object
          nullable: true
          additionalProperties: { type: string }
    Money:
      type: integer
      format: int64
      minimum: 0
      description: Amount in paise. 25000 = ₹250.00. Never a decimal.
    OrderStatus:
      type: string
      enum:
        [
          CREATED,
          PAYMENT_PENDING,
          PAYMENT_CONFIRMED,
          DISPATCHED,
          ACKNOWLEDGED,
          PREPARING,
          READY,
          COMPLETED,
          PAYMENT_FAILED,
          DISPATCH_FAILED,
          REJECTED,
          CANCELLED,
          REFUND_PENDING,
          REFUNDED,
          REFUND_FAILED,
          RECONCILIATION_REQUIRED,
        ]
...
    SettlementMode:
      type: string
      enum: [PLATFORM_COLLECT, VENDOR_DIRECT]
      description: |
        PRD §4. Never exposed in the customer PWA (PAY-MODE-07). Present here
        only on staff and admin surfaces.
...
    RejectionReason:
      type: string
      enum:
        [
          ITEM_OUT_OF_STOCK,
          KITCHEN_OVERLOADED,
          EQUIPMENT_FAILURE,
          VENDOR_CLOSING,
          INGREDIENT_UNAVAILABLE,
          PRICE_OR_MENU_ERROR,
          CUSTOMER_REQUEST,
          DUPLICATE_ORDER,
          OTHER,
        ]
    UnavailableReason:
      type: string
      enum: [OUTSIDE_HOURS, TEMPORARILY_CLOSED, DEVICE_OFFLINE, DISPATCH_BLOCKED, SUSPENDED]
...
    VendorSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        cuisine:
          type: array
          items: { type: string }
        imageUrl: { type: string, nullable: true }
        orderable: { type: boolean }
        unavailableReason:
          oneOf:
            - { $ref: '#/components/schemas/UnavailableReason' }
            - { type: 'null' }
        estimatedPrepMinutes: { type: integer }
...
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        publicOrderNumber: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        foodCourtId: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        vendorName: { type: string }
        tableLabel: { type: string }
        items:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              quantity: { type: integer }
              lineTotalPaise: { $ref: '#/components/schemas/Money' }
              options:
                type: array
                items: { type: string }
              instructions: { type: string, nullable: true }
              isRejected: { type: boolean }
        totalPayablePaise: { $ref: '#/components/schemas/Money' }
        rejectionReason:
          oneOf:
            - { $ref: '#/components/schemas/RejectionReason' }
            - { type: 'null' }
        refund:
          oneOf:
            - { $ref: '#/components/schemas/Refund' }
            - { type: 'null' }
        creditIssued:
          oneOf:
            - { $ref: '#/components/schemas/PlatformCredit' }
            - { type: 'null' }
        cancellationOffered:
          type: boolean
          description: True once the escalation ladder reaches 180s (PRD ESC-03).
        estimatedReadyAt: { type: string, format: date-time, nullable: true }
        timeline:
          type: array
          items:
            type: object
            properties:
              status: { $ref: '#/components/schemas/OrderStatus' }
              at: { type: string, format: date-time }
              actorType: { type: string }
        createdAt: { type: string, format: date-time }
...
    Refund:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        kind: { type: string, enum: [FULL, PARTIAL] }
        amountPaise: { $ref: '#/components/schemas/Money' }
        status: { type: string, enum: [REQUESTED, PENDING, SUCCEEDED, FAILED, ABANDONED] }
        customerMessage:
          type: string
          description: |
            Explicit customer-facing status. The words "payment lost" or an
            unexplained failure never appear (PRD CUS-TRK-05).
        expectedByAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    PlatformCredit:
      type: object
      description: |
        PRD REF-05. Issued instantly at rejection so the customer is not left
        waiting on a bank refund. Mutually exclusive with refund-to-source,
        enforced atomically.
      properties:
        id: { type: string, format: uuid }
        amountPaise: { $ref: '#/components/schemas/Money' }
        status: { type: string, enum: [ISSUED, CONSUMED, EXPIRED, VOIDED] }
        expiresAt: { type: string, format: date-time }
...
    VendorInput:
      type: object
      required: [foodCourtId, name, settlementMode]
      properties:
        foodCourtId: { type: string, format: uuid }
        name: { type: string }
        legalName: { type: string, nullable: true }
        cuisine:
          type: array
          items: { type: string }
        settlementMode: { $ref: '#/components/schemas/SettlementMode' }
        estimatedPrepMinutes: { type: integer, minimum: 1, maximum: 180, default: 10 }
    Vendor:
      allOf:
        - { $ref: '#/components/schemas/VendorInput' }
        - type: object
          properties:
            id: { type: string, format: uuid }
            status: { type: string }
            providerLinkedAccountId: { type: string, nullable: true }
...